Populate .. contents:: directive with actual contents
The .. contents:: directive had the option :local: but had no
subsections in its section, thus the table of contents was empty.
Remove the .. contents:: directive section and merge it with its super
section which contains further subsections so that some content is
produced.  Also suppress backlinks to not make the subsections headers
a link back to this .. contents:: directive.  The link markup destroys
the header markup and does not look pretty.